backend/pyproject.toml
[project]
name = "dog-eared-backend"
version = "0.1.0"
description = "Dog-Eared backend (FastAPI)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "python-dotenv>=1.0.1",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.1.3",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "rapidfuzz>=3.9.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
